remove outdated files, clean up checkpoint dir/wandb logging (#31)
Former-commit-id: 020a21d4dbcc76954708cce5e3c993a952711d3d
