+++ amptorch/model.py
@@ -1,39 +1,13 @@
-"""NN_model.py: Constructs a model consisting of element specific Neural
-Networks as understood from Behler and Parrinello's works. A model instance is
-constructed based off the unique number of atoms in the dataset."""
-
-import sys
-from collections import defaultdict
 import torch
 import torch.nn as nn
-from torch.nn import Tanh, Softplus, LeakyReLU
-from torch.nn import init
-from torch.nn.init import xavier_uniform_, kaiming_uniform_
 from torch.autograd import grad
-from amptorch.utils import Logger
+from torch.nn import Tanh
+from torch_scatter import scatter
 
-__author__ = "Muhammed Shuaibi"
-__email__ = "mshuaibi@andrew.cmu.edu"
 
 class MLP(nn.Module):
-    """Constructs a fully connected neural network model to be utilized for
-    each element type'''
-
-    Arguments:
-        n_input_nodes: Number of input nodes (Default=20 using BP SF)
-        n_output_nodes: Number of output nodes (Default=1)
-        n_layers: Total number of layers in the neural network
-        n_hidden_size: Number of neurons within each hidden layer
-        activation: Activation function to be utilized. (Default=Tanh())
-    """
-
     def __init__(
-        self,
-        n_input_nodes,
-        n_layers,
-        n_hidden_size,
-        activation,
-        n_output_nodes=1,
+        self, n_input_nodes, n_layers, n_hidden_size, activation, n_output_nodes=1
     ):
         super(MLP, self).__init__()
         if isinstance(n_hidden_size, int):
@@ -47,209 +21,115 @@ class MLP(nn.Module):
         layers.append(nn.Linear(self.n_neurons[-2], self.n_neurons[-1]))
         self.model_net = nn.Sequential(*layers)
 
-    def forward(self, inputs):
-        """Feeds data forward in the neural network
+        # TODO: identify optimal initialization scheme
+        # self.reset_parameters()
 
-        Arguments:
-            inputs (torch.Tensor): NN inputs
-        """
+    def reset_parameters(self):
+        for m in self.model_net:
+            if isinstance(m, torch.nn.Linear):
+                torch.nn.init.xavier_uniform_(m.weight)
+                m.bias.data.fill_(0)
+
+    def forward(self, inputs):
         return self.model_net(inputs)
 
 
-class FullNN(nn.Module):
-    """Combines element specific NNs into a model to predict energy of a given
-    structure
+class ElementMask(nn.Module):
+    def __init__(self, elements):
+        super(ElementMask, self).__init__()
+        nelems = len(elements)
+        weights = torch.zeros(100, nelems)
+        weights[elements, range(nelems)] = 1.0
 
-    """
+        self.mask = nn.Embedding(100, nelems)
+        self.mask.weight.data = weights
 
+    def forward(self, atomic_numbers):
+        return self.mask(atomic_numbers)
+
+
+class BPNN(nn.Module):
     def __init__(
-        self, unique_atoms, architecture, device, forcetraining,
-        activation=Tanh, require_grd=True
+        self,
+        elements,
+        input_dim,
+        num_nodes,
+        num_layers,
+        get_forces=True,
+        activation=Tanh,
     ):
-        super(FullNN, self).__init__()
-        self.device = device
-        self.req_grad = require_grd
-        self.forcetraining = forcetraining
-        self.architecture = architecture
+        super(BPNN, self).__init__()
+        self.get_forces = get_forces
         self.activation_fn = activation
 
-        input_length = architecture[0]
-        n_layers = architecture[1]
-        n_hidden_size = architecture[2]
-        self.elementwise_models = nn.ModuleDict()
-        for element in unique_atoms:
-            self.elementwise_models[element] = MLP(
-                n_input_nodes=input_length,
-                n_layers=n_layers,
-                n_hidden_size=n_hidden_size,
-                activation=activation,
+        n_elements = len(elements)
+        self.elementwise_models = nn.ModuleList()
+        for element in range(n_elements):
+            self.elementwise_models.append(
+                MLP(
+                    n_input_nodes=input_dim,
+                    n_layers=num_layers,
+                    n_hidden_size=num_nodes,
+                    activation=activation,
+                )
             )
 
-    def forward(self, inputs):
-        """Forward pass through the model - predicting energy and forces
-        accordingly.
+        self.element_mask = ElementMask(elements)
 
-        N - Number of training images
-        Q - Atoms in batch
-        P - Length of fingerprint"""
-
+    def forward(self, batch):
         with torch.enable_grad():
-            input_data = inputs[0]
-            batch_size = inputs[1]
-            batch_elements = inputs[2]
-            rearange = inputs[-1]
-            if self.device == 'cpu':
-                energy_pred = torch.zeros(batch_size, 1).to(self.device)
-            else:
-                energy_pred = torch.cuda.FloatTensor(batch_size, 1)
-            force_pred = torch.tensor([])
-            # Constructs an Nx1 empty tensor to store element energy contributions
-            if self.forcetraining:
-                fprimes = inputs[-2]
-                dE_dFP = torch.tensor([]).to(self.device)
-                idx = torch.tensor([]).to(self.device)
-            for index, element in enumerate(batch_elements):
-                model_inputs = input_data[element][0]
-                model_inputs.requires_grad = True
-                contribution_index = torch.tensor(input_data[element][1]).to(self.device)
-                atomwise_outputs = self.elementwise_models[element].forward(model_inputs)
-                energy_pred.index_add_(0, contribution_index, atomwise_outputs)
-                if self.forcetraining:
-                    gradients = grad(
-                        energy_pred,
-                        model_inputs,
-                        grad_outputs=torch.ones_like(energy_pred),
-                        create_graph=True,
-                    )[0]
-                    dE_dFP = torch.cat((dE_dFP, gradients))
-                    idx = torch.cat((idx, contribution_index.float()))
-            if self.forcetraining:
-                """Constructs a 1xPQ tensor that contains the derivatives with respect to
-                each atom's fingerprint"""
-                boolean = idx[:, None] == torch.unique(idx)
-                ordered_idx = torch.nonzero(boolean.t(), as_tuple=False)[:, -1]
-                dE_dFP = torch.index_select(dE_dFP, 0, ordered_idx)
-                dE_dFP = torch.index_select(dE_dFP, 0, rearange).reshape(1, -1)
-                """Sparse multiplication requires the first matrix to be
-                sparse.
-                Multiplies a 3QxPQ tensor with a PQx1 tensor to return a 3Qx1 tensor
-                containing the x,y,z directional forces for each atom"""
-                force_pred = -1 * torch.sparse.mm(fprimes.t(), dE_dFP.t())
-                """Reshapes the force tensor into a Qx3 matrix containing all the force
-                predictions in the same order and shape as the target forces calculated
-                from AMP."""
-                force_pred = force_pred.reshape(-1, 3)
-        return energy_pred, force_pred
+            atomic_numbers = batch.atomic_numbers
+            fingerprints = batch.fingerprint.float()
+            fingerprints.requires_grad = True
+            image_idx = batch.image_idx
+            sorted_image_idx = torch.unique_consecutive(image_idx)
+            mask = self.element_mask(atomic_numbers)
+            o = torch.sum(
+                mask
+                * torch.cat(
+                    [net(fingerprints) for net in self.elementwise_models], dim=1
+                ),
+                dim=1,
+            )
+            energy = scatter(o, image_idx, dim=0)[sorted_image_idx]
+
+            if self.get_forces:
+                gradients = grad(
+                    energy,
+                    fingerprints,
+                    grad_outputs=torch.ones_like(energy),
+                    create_graph=True,
+                )[0].view(1, -1)
+
+                forces = -1 * torch.sparse.mm(batch.fprimes.t(), gradients.t()).view(
+                    -1, 3
+                )
+            return energy, forces
+
+    @property
+    def num_params(self):
+        return sum(p.numel() for p in self.parameters())
+
 
 class CustomMSELoss(nn.Module):
-    """Custom loss function to be optimized by the regression. Includes aotmic
-    energy and force contributions.
-
-    Eq. (26) in A. Khorshidi, A.A. Peterson /
-    Computer Physics Communications 207 (2016) 310-324"""
-
     def __init__(self, force_coefficient=0):
         super(CustomMSELoss, self).__init__()
         self.alpha = force_coefficient
 
-    def forward(
-        self,
-        prediction,
-        target):
+    def forward(self, prediction, target):
 
         energy_pred = prediction[0]
-        energy_targets_per_atom = target[0]
-        num_atoms = target[1]
-        MSE_loss = nn.MSELoss(reduction="sum")
-        energy_pred_per_atom = torch.div(energy_pred, num_atoms)
-        energy_loss = MSE_loss(energy_pred_per_atom, energy_targets_per_atom)
+        energy_target = target[0]
+        MSE_loss = nn.MSELoss()
+        energy_loss = MSE_loss(energy_pred, energy_target)
 
         if self.alpha > 0:
             force_pred = prediction[1]
+            force_target = target[1]
             if force_pred.nelement() == 0:
-                raise Exception('Force training disabled. Set force_coefficient to 0')
-            force_targets_per_atom = target[-1]
-            num_atoms_extended = torch.cat([idx.repeat(int(idx)) for idx in num_atoms])
-            num_atoms_extended = torch.sqrt(
-                num_atoms_extended.reshape(-1, 1)
-            )
-            force_pred_per_atom = torch.div(force_pred, num_atoms_extended)
-            force_targets_per_atom = force_targets_per_atom*num_atoms_extended
-            force_loss = (self.alpha / 3) * MSE_loss(
-                force_pred_per_atom, force_targets_per_atom
-            )
+                raise Exception("Force training disabled. Set force_coefficient to 0")
+            force_loss = MSE_loss(force_pred, force_target)
             loss = 0.5 * (energy_loss + force_loss)
         else:
             loss = 0.5 * energy_loss
-        return loss
-
-class MAELoss(nn.Module):
-    def __init__(self, force_coefficient=0):
-        super(MAELoss, self).__init__()
-        self.alpha = force_coefficient
-
-    def forward(
-        self,
-        prediction,
-        target):
-
-        energy_pred = prediction[0]
-        energy_targets = target[0]
-        num_atoms = target[1]
-        MAE_loss = nn.L1Loss(reduction="sum")
-        energy_per_atom = torch.div(energy_pred, num_atoms)
-        targets_per_atom = torch.div(energy_targets, num_atoms)
-        energy_loss = MAE_loss(energy_per_atom, targets_per_atom)
-
-        if self.alpha > 0:
-            force_pred = prediction[1]
-            if force_pred.nelement() == 0:
-                raise Exception('Force training disabled. Set force_coefficient to 0')
-            force_targets = target[-1]
-            num_atoms_force = torch.cat([idx.repeat(int(idx)) for idx in num_atoms])
-            num_atoms_force = num_atoms_force.reshape(len(num_atoms_force), 1)
-            force_pred_per_atom = torch.div(force_pred, num_atoms_force)
-            force_targets_per_atom = torch.div(force_targets, num_atoms_force)
-            force_loss = (self.alpha / 3) * MAE_loss(
-                force_pred_per_atom, force_targets_per_atom
-            )
-            loss = 0.5 * (energy_loss + force_loss)
-        else:
-            loss = 0.5 * energy_loss
-        return loss
-
-class HuberLoss(nn.Module):
-    def __init__(self, force_coefficient=0):
-        super(HuberLoss, self).__init__()
-        self.alpha = force_coefficient
-
-    def forward(
-        self,
-        prediction,
-        target):
-
-        energy_pred = prediction[0]
-        energy_targets = target[0]
-        num_atoms = target[1]
-        huber_loss = nn.SmoothL1Loss(reduction="none")
-        energy_per_atom = torch.div(energy_pred, num_atoms)
-        targets_per_atom = torch.div(energy_targets, num_atoms)
-        energy_loss = huber_loss(energy_per_atom, targets_per_atom)
-        energy_loss = torch.sum(energy_loss)
-
-        if self.alpha > 0:
-            force_pred = prediction[1]
-            if force_pred.nelement() == 0:
-                raise Exception('Force training disabled. Set force_coefficient to 0')
-            force_targets = target[-1]
-            num_atoms_force = torch.cat([idx.repeat(int(idx)) for idx in num_atoms]).reshape(-1, 1)
-            force_pred_per_atom = torch.div(force_pred, num_atoms_force)
-            force_targets_per_atom = torch.div(force_targets, num_atoms_force)
-            force_loss = (self.alpha / 3) * huber_loss(
-                force_pred_per_atom, force_targets_per_atom
-            )
-            force_loss = torch.sum(force_loss)
-            loss = energy_loss + force_loss
-        else:
-            loss = energy_loss
         return loss
